refactor: Remove Context API logic To reuse api class logic in next.js api folder which supports server
