+++ .env
@@ -1,0 +1,1 @@
+BASE_URL = http://ownus:3030/api
